refactor(index): move all imports and modules to a separate module.ts file. The motivation behind this change is those that wish to use the exports and create their own NgModules out of our components can do so without having an import explosion and including all of Ionic
